Compile fixes for gcc -Wall

+++ glslang/MachineIndependent/Scan.h
@@ -172,7 +172,7 @@ protected:
     void advance()
     {
         ++currentChar;
-        if (currentChar >= static_cast<int>(lengths[currentSource])) {
+        if (currentChar >= static_cast<size_t>(lengths[currentSource])) {
             ++currentSource;
             if (currentSource < numSources) {
                 loc[currentSource].string = loc[currentSource - 1].string + 1;
